Add MOV Immediate support
I also added setCarry() and clearCarry() methods to the pinkySimBase
class so that the initial state of the carry flag can be set.  This was
added since MOV actually just maintains the carry flag state but the
setExpectedAPSRflags() method sets it to the opposite of the expected
post condition to make sure that it gets flipped.  Now a test can set
the expected value and then explicitly set the precondition of the
flag to make sure that MOV doesn't modify it.