Verify APSR unused flags aren't modified
Some instructions such as AND shouldn't touch all of the APSR flags (in
particular the oVerflow flag.)  During setup() of each test, I now
randomly initialize each APSR flag.  If the test doesn't
explicitly verify a flag through the use of the setExpectedAPSRflags()
method then it has a 50% chance of being set before running the step
instruction and is verified to remain the same after the step.