Modify coverage script to show lines missed
I currently have a few lines which are always missed.  I usually added
the lines to keep the compiler happy (missing cases in switch statements
or all code paths having return value, etc) but I expect many of them
to be covered in the future.  I modified the script to use grep to show
me the missed lines so that I can quickly see if they are the ones I
expect.