Update shift related comments
The architecture manual will sometimes show a Shift_C() call which is
hardcoded for ARMv6-m scenarios to be a LSL of #0.  This means that the
shift has no impact on the value.  I previously had a comment about
this fact for instruction like AND, EOR, ORR, etc.  I added some text
to this comment which indicated that there is a side effect from this
call which causes the state of the APSR carry flag to be maintained.
If I remove this extraneous code in the future, I should remove the
code below which manipulates the carry flag as well.