Add an unpredictable instruction + fix addRegisterT2() decode.
The special data and branch exchange class of 16-bit instructions has
one encoding which is specified as being unpredictable.  Added tests to
make sure that pinkySim returns PINKYSIM_STEP_UNPREDICTABLE for this
encoding.

I also fixed a decoding bitmasking error in
specialDataAndBranchExchange() which sent the wrong bit patterns to be
processed as addRegisterT2().