Recent MRI fixes and associated thunks updates
Recent fixes to MRI allow enabling of the BL tests which verify range
limits (resulting in the PC pointing to invalid memory).

MRI no longer results in Hard Faults when attempting to step over a SVC
instruction.  The realarm program was modified to include a SVC_Handler
which just returns and the thunking version of pinkySimStep() now knows
how to detect that this handler has been entered and then step out of
it to more closely resemble what the simulator does.