Add MemorySim_CreateRegionsFromFlashImage()
This new API will take a .bin file meant to be loaded into the FLASH of
a Cortex-M processor and load it into the simulated FLASH starting at
address 0x0000000.  This read-only FLASH region will just be large
enough to contain the contents of the .bin file.  The API will then
query the first word in the image, the initial SP value, and use it
to create an initial RAM region.  This region will end 1 byte before
the initial SP value and start at the intial SP value ANDed with
0xF0000000.

This will allow much of my future testing with the simulator to only
need specifying of memory ranges in the linker script used for the
generation of test images.  pinkySim will just do what it needs to for
loading everything into the correct locations.