Start of libmri4sim
This is the start of the library which will integrate MRI debug stub
support into pinkySim.  This code just provides the start of the
Platform layer required by libmricore.

I also updated the makefile to treat non-file targets such as
GCOV_LIBPINKYSIM as .PHONY targets so that files existing with the same
name don't stop these targets from being built.

I updated the MRI subproject to include a fix which only defines the
__debugbreak() macro if not already defined.  This version was for
ARM anyway so it doesn't help when running pinkySim code on the host.