Start of mri4simRun() implementation/test
Added more communication APIs to the IComm interface and renamed
shouldExit() to shouldStopRun().

Added a mockIComm mock to implement a test IComm interface to be used
by the mri4sim unit tests to queue up receive data and verify transmit
data.

Provided actual implementations for more of the Platform_* APIs in
mri4sim.c:
* Platform_GetPacketBuffer()
* Platform_GetPacketBufferSize()
* Platform_EnteringDebugger()
* Platform_CommHasReceiveData()
* Platform_CommReceiveChar()
* Platform_CommSendChar()
* Platform_DetermineCauseOfException()
* Platform_AdvanceProgramCounterToNextInstruction()
* Platform_WasProgramCounterModifiedByUser()
* Platform_WriteTResponseRegistersToBuffer()
* Platform_TypeOfCurrentInstruction()

Added a mri4simBaseTest.h which can be used as the base class for
mri4sim unit tests.  It takes care of setup/cleanup, and emitting
Thumb instructions.