Add pinkySim application
Now builds a pinkySim application which actually pulls all of the
modules that I have been working on into an actual functional
application that can actually run and execute ARMv6-M samples.

The resulting application is called pinkySim and once built, will be
found in the root directory of the project.