Support main() input arguments and return value
pinkySim now accepts optional arguments on the command line after the
name of the image filename and passes them into the main() of the
program running under simulation.  It also takes the return value from
main() of the simulated program and returns it back to the shell which
launched pinkySim itself.

I also added a CommandLine sample which demos this new feature.