32-byte align top of stack
I was hitting issues where GDB would try to read past the end of my
stack.  It would just go and grab a bunch of data off of the stack when
it attempted to dump locals in Reset_Handler().  It appears to assume
that it can safely round up to the nearest 32-byte boundary when
performing this read.  Once I 32-byte aligned the size of my RAM
section in samples/pinkySim.ld, this problem went way.  32-bytes are
the finest granularity for MPU regions so maybe the assumption is
related to that fact.