Add MemorySim_MapSimulatedAddressToHostAddress*()
Added the following two new APIs to the MemorySim module:
    MemorySim_MapSimulatedAddressToHostAddressForWrite()
    MemorySim_MapSimulatedAddressToHostAddressForRead()
These APIs allow the simulator to take an address used by code running
within the simulator and map it to the actual memory in the host system
running the simulator. This is useful for implementing newlib syscall
thunks in the simulator when addresses for read and writes are passed
in from the code running on the simulator.  The simulator can call the
corresponding (read/write) mapping API to get a valid address for the
the simulator and use it when calling the desired host Posix function
on behalf of the simulated program.