Extend read()/write() mocks
I extended the write mock to track writes to files other than stdout
stderr.  I also added the ability to have the write() mock return an
error back to the caller and set the global errno.

I extended the read mock to not return an error for accessing any file
other than stdin but instead returning data from the read buffer just
as it already did for stdin reads.  This allows mocking of read() calls
performed for any file handle.