Update newlib semihost support
Previously most of the newlib semihost handlers in pinkySim would pass
syscall requests onto GDB.  This meant that GDB needed to be connected
to the simulator to perform any file I/O.  This commit modifies things
so that most newlib syscalls will be handled by calling Posix APIs
directly on the host rather than marshalling the requests over to GDB.
The only exception is to send stdout, stdin, and stderr read/writes to
GDB if it is attached.

I did modify the handling of fstat() and stat() calls a bit.  Now a
minimal stat structure is filled in by pinkySim (it just contains
mode, size, blocks, and blksize fields).  These fields are then copied
into the awaiting newlib stat structure by code in NewlibRetarget.c
which is linked into the simulated program.  This way the host and
target can share a stat structure definition even though each of theirs
will contain differences (OS X and newlib for example).