Enable GCOV code coverage in samples
The samples now create two output files for each sample. A normal one
as it did before and another with a _gcov.bin suffix. This latter one
will record code coverage data while running under the simulator and
write out the results just before it exits.  This takes advantage of
the newlib semihosting in pinkySim to allow the gcov instrumentation
to open files on the host and write out the code coverage results.

The run-gcov.sh bash script can be executed (tested on OS X only) to
run arm-none-eabi-gcov over the results and extract summary and
detailed code coverage results for a particular sample.  The sample
must be manually executed under the simulator by the user before
running this script to ascertain the code coverage from that run.