Don't return from within __try clause
It is ok for a function called from within a __try clause to return
since it will return to the __try clause but calling return
directly from within the __try clause will exit the clause without
cleaning up the registered exception handler.  This lead to leaked
handlers in the linked list of exception handlers.