Add ElfLines modue
You give this module the filename of your .ELF executable file and it
runs arm-none-eabi-readelf on it to extract the source line to device
memory address mappings.  This will be used along with the read
counting functionality that was recently added to the MemorySim module
to produce human readable code coverage results.