Handle fragmented source files in ElfLines
Previously the doesCurrentSourceLineMatchCurrentElfLine() routine only
compared the filename pointers rather than performing a string compare.
This failed to work as expected if the elf output fragmented the
information for a particular source file.  This would generate two
different filename strings in the ElfLines output but they still refer
to the same file and are sorted into the correct location.